[#395] REFACTOR: chat bubbles 로직 분리
